+++ slim/endpoints/localidades.php
@@ -1,5 +1,4 @@
 <?php
-// Container Slim?
 // ¿Vamos a ver todo lo que es la configuracion APACHE para un servidor? Yo no termino de entender que es lo que nos está permitiendo tener un servidor propio local desde Docker. Osea por ej yo se que con react haces un npm run dev y te levanta el localHost con los archivos pero con docker nose como funciona.
 
 namespace Endpoints;
@@ -12,27 +11,32 @@ use Psr\Http\Message\ServerRequestInterface as Request;
 
 class Localidades extends Endpoint
 {
-
-    public function crear(Request $request, Response $response) // ¿Tengo que validar que la localidad no tenga numeros en su nombre? DUDA
+    public function crear(Request $request, Response $response)
     {
-
-        $msg = '';
         try {
 
             $connection = $this->getConnection();
-
             $nuevaLocalidad = $request->getParsedBody();
 
-            $tablaLocalidades = $connection->prepare('SELECT * FROM localidades WHERE nombre = :nuevoNombre');
-            $tablaLocalidades->execute([':nuevoNombre' => $nuevaLocalidad]);
+            $tablaLocalidades = $connection->prepare('SELECT nombre FROM localidades WHERE nombre = :nuevoNombre');
+            $tablaLocalidades->execute([':nuevoNombre' => $nuevaLocalidad['nombre']]);
             $existe = $tablaLocalidades->fetchColumn();
 
-            if (!$existe) {
+            if (!$existe && isset($nuevaLocalidad)) {
                 $tablaLocalidades = $connection->prepare('INSERT INTO localidades (nombre) VALUES (:nuevoNombre)');
-                $tablaLocalidades->execute([':nuevoNombre' => $nuevaLocalidad]);
-                $msg = 'La localidad ha sido agregada correctamente';
+                $tablaLocalidades->execute([':nuevoNombre' => $nuevaLocalidad['nombre']]);
+                $this->data['Status'] = 'Success';
+                $this->data['Mensaje'] = 'La localidad ha sido agregada correctamente.';
+                $this->data['Data'] = $nuevaLocalidad;
+                $this->data['Codigo'] = '200';
             } else {
-                $msg = 'Error: La localidad ingresada ya existe en la base de datos.';
+                $this->data['Status'] = 'Fail';
+                $this->data['Data'] = $nuevaLocalidad;
+                $this->data['Codigo'] = '400';
+                if ($existe)
+                    $this->data['Mensaje'] = 'La localidad ingresada ya se encuentra en la base da datos.';
+                else
+                    $this->data['Mensaje'] = 'No se ha ingresado ningún nombre de localidad';
             }
 
         } catch (Exception $e) {
@@ -55,24 +59,47 @@ class Localidades extends Endpoint
             $localidad->execute([':num' => $id]);
             $existeID = $localidad->fetchColumn();
 
-            if ($existeID) {
-                $cambioNombre = $request->getParsedBody();
+            // No se puede editar una localidad si ya esta asociada a la tabla de propiedades
+            $propiedad = $connection->prepare('SELECT * FROM propiedades WHERE localidad_id = :num');
+            $propiedad->execute([':num' => $id]);
+            $existePropiedad = $propiedad->fetch(PDO::FETCH_ASSOC);
 
-                $localidad = $connection->prepare('SELECT nombre FROM localidades WHERE nombre = :nuevoNombre and id <> :num'); // Verifica si el nuevo nombre se encuentra usado ya en otro ID existente.
-                $localidad->execute([':nuevoNombre' => $cambioNombre]);
-                $existeNombre = $localidad->fetchColumn();
+            if (!$existePropiedad) {
+                if ($existeID) {
+                    $cambioNombre = $request->getParsedBody();
 
-                if (!$existeNombre) {
-                    $localidad = $connection->prepare('UPDATE localidades SET nombre = :nuevoNombre WHERE id = :num');
-                    $localidad->execute([':num' => $id, ':nuevoNombre' => $cambioNombre]);
-                    $msg = 'Cambio de nombre de localidad realizado.';
-                } else
-                    $msg = 'Error: El nombre ya existe.';
-            } else
-                $msg = [
-                    'Error ' => 'La localidad correspondiente al ID ingresado no existe.',
-                    'ID buscado ' => $id,
-                ];
+                    $localidad = $connection->prepare('SELECT nombre FROM localidades WHERE nombre = :nuevoNombre and id <> :num'); // Verifica si el nuevo nombre se encuentra usado ya en otro ID existente.
+                    $localidad->execute([':nuevoNombre' => $cambioNombre]);
+                    $existeNombre = $localidad->fetchColumn();
+
+                    if (!$existeNombre && isset($cambioNombre)) {
+                        $localidad = $connection->prepare('UPDATE localidades SET nombre = :nuevoNombre WHERE id = :num');
+                        $localidad->execute([':num' => $id, ':nuevoNombre' => $cambioNombre]);
+                        $this->data['Status'] = 'Success';
+                        $this->data['Mensaje'] = 'La localidad ha sido agregada correctamente.';
+                        $this->data['Data'] = $cambioNombre;
+                        $this->data['Codigo'] = '200';
+                    } else {
+                        $this->data['Status'] = 'Fail';
+                        $this->data['Data'] = $cambioNombre;
+                        $this->data['Codigo'] = '400';
+                        if ($existeNombre)
+                            $this->data['Mensaje'] = 'La localidad ingresada ya se encuentra en la base da datos.';
+                        else
+                            $this->data['Mensaje'] = 'No se ha ingresado ningún nombre de localidad';
+                    }
+                } else {
+                    $this->data['Status'] = 'Fail';
+                    $this->data['Mensaje'] = 'La localidad correspondiente al ID ingresado no existe.';
+                    $this->data['Data'] = $id;
+                    $this->data['Codigo'] = '400';
+                }
+            } else {
+                $this->data['Status'] = 'Fail';
+                $this->data['Mensaje'] = 'La localidad no se puede editar porque ya se encuentra asociada a una propiedad.';
+                $this->data['Data'] = $existePropiedad;
+                $this->data['Codigo'] = '400';
+            }
 
         } catch (Exception $e) {
             $this->data['Status'] = 'Throw Exception';
@@ -90,20 +117,35 @@ class Localidades extends Endpoint
             $id = $args['id'];
             $connection = $this->getConnection();
 
+            // No se puede eliminar una localidad si ya esta asociada a la tabla de propiedades
+            $propiedad = $connection->prepare('SELECT * FROM propiedades WHERE localidad_id = :num');
+            $propiedad->execute([':num' => $id]);
+            $existePropiedad = $propiedad->fetch(PDO::FETCH_ASSOC);
 
-            $localidad = $connection->prepare('SELECT * FROM localidades WHERE id = :num');
-            $localidad->execute([':num' => $id]);
-            $existeID = $localidad->fetchColumn();
-
-            if ($existeID) {
-                $localidad = $connection->prepare('DELETE FROM localidades WHERE id = :num');
+            if (!$existePropiedad) {
+                $localidad = $connection->prepare('SELECT * FROM localidades WHERE id = :num');
                 $localidad->execute([':num' => $id]);
-                $msg = 'La localidad buscada ha sido eliminada correctamente.';
-            } else
-                $msg = [
-                    'Error ' => 'La localidad correspondiente al ID ingresado no existe.',
-                    'ID buscado ' => $id,
-                ];
+                $existeID = $localidad->fetch(PDO::FETCH_ASSOC);
+
+                if ($existeID) {
+                    $this->data['Data'] = $existeID;
+                    $localidad = $connection->prepare('DELETE FROM localidades WHERE id = :num');
+                    $localidad->execute([':num' => $id]);
+                    $this->data['Status'] = 'Success';
+                    $this->data['Mensaje'] = 'La localidad ha sido eliminada correctamente.';
+                    $this->data['Codigo'] = '200';
+                } else {
+                    $this->data['Status'] = 'Fail';
+                    $this->data['Mensaje'] = 'La localidad correspondiente al ID ingresado no existe.';
+                    $this->data['Data'] = $id;
+                    $this->data['Codigo'] = '400';
+                }
+            } else {
+                $this->data['Status'] = 'Fail';
+                $this->data['Mensaje'] = 'La localidad no se puede eliminar porque ya se encuentra asociada a una propiedad.';
+                $this->data['Data'] = $existePropiedad;
+                $this->data['Codigo'] = '400';
+            }
 
         } catch (Exception $e) {
             $this->data['Status'] = 'Throw Exception';
